Changed max yintercept to 30 on line 266
Student commented that using 40 as a maximum value does not make much sense when the aim is to highlight where the upper limit should be (i.e., 30). I would also argue 0 is not quite right since the start of the chapter explains minus values are possible, but I don't know what you would set for this instead as we don't know an absolute lower limit.